Move atom class to a separate module and add Hydrogen atom subclass.
